Treat handle mismatch as regular test failure
When the model and the system under test return a different number of
handles, QSM currently fails with an error message that suggests that
this is an indication that there is a bug in the test setup. This does
not however need to be the case.

Consider testing an SQL database which generates primary keys
automatically on doing `INSERT`. We might want to allow the system under
test to use different primary keys from the model in such a scenario;
after all, the exact mechanism by which the database picks primary keys
is not relevant.

To model this in QSM, we have to use references for primary keys. But
now consider doing a `SELECT`. It's entirely conceivable that due to a
bug in the code somewhere (either in the model or in the system under
test) the number of rows returned might be different in the model than
in the system under test. When this happens, however, the model will
also return a different number of _handles_.

Of course, we cannot continue at this point -- but QSM _anyway_ does not
continue after an error. This commit does not change that behaviour;
instead, it merely changes the error message to avoid suggesting that
this a bug in the test setup.