Remove the tree-diff dependency.
Vendor/copy in the relevant bits that we need from the 0.0.2.1 version
instead. The reason for this is that after that version the license
was changed from BSD to GPL and pinning the dependency to that version
doesn't compile with newer GHC versions, by inlining tree-diff we are
in control of its dependecy bounds (and can thus make it compile with
newer versions of GHC).